Refactoring. Spatial transforms moved to geometric files. Better type hints
