+++ albumentations/augmentations/dropout/coarse_dropout.py
@@ -1,5 +1,4 @@
 import random
-import typing
 from typing import Iterable, List, Optional, Sequence, Tuple, Union
 
 import numpy as np
