+++ albumentations/augmentations/functional.py
@@ -1,6 +1,5 @@
 from __future__ import division
 
-import math
 from functools import wraps
 from itertools import product
 from typing import Optional, Sequence, Union
@@ -27,10 +26,6 @@ __all__ = [
     "adjust_hue_torchvision",
     "adjust_saturation_torchvision",
     "angle_2pi_range",
-    "bbox_flip",
-    "bbox_hflip",
-    "bbox_transpose",
-    "bbox_vflip",
     "blur",
     "brightness_contrast_adjust",
     "channel_shuffle",
@@ -39,7 +34,6 @@ __all__ = [
     "clipped",
     "convolve",
     "downscale",
-    "elastic_transform_approx",
     "equalize",
     "fancy_pca",
     "from_float",
@@ -49,19 +43,12 @@ __all__ = [
     "get_num_channels",
     "get_opencv_dtype_from_numpy",
     "glass_blur",
-    "grid_distortion",
-    "hflip",
-    "hflip_cv2",
     "image_compression",
     "invert",
     "is_grayscale_image",
     "is_multispectral_image",
     "is_rgb_image",
     "iso_noise",
-    "keypoint_flip",
-    "keypoint_hflip",
-    "keypoint_transpose",
-    "keypoint_vflip",
     "linear_transformation_rgb",
     "median_blur",
     "move_tone_curve",
@@ -69,14 +56,9 @@ __all__ = [
     "non_rgb_warning",
     "noop",
     "normalize",
-    "optical_distortion",
-    "pad",
-    "pad_with_params",
     "posterize",
     "preserve_channel_dim",
     "preserve_shape",
-    "random_flip",
-    "rot90",
     "shift_hsv",
     "shift_rgb",
     "solarize",
@@ -84,9 +66,7 @@ __all__ = [
     "swap_tiles_on_image",
     "to_float",
     "to_gray",
-    "transpose",
     "unsharp_mask",
-    "vflip",
 ]
 
 MAX_VALUES_BY_DTYPE = {
@@ -216,32 +196,6 @@ def non_rgb_warning(image):
             message += "\nThis transformation cannot be applied to multi-spectral images"
 
         raise ValueError(message)
-
-
-def vflip(img):
-    return np.ascontiguousarray(img[::-1, ...])
-
-
-def hflip(img):
-    return np.ascontiguousarray(img[:, ::-1, ...])
-
-
-def hflip_cv2(img):
-    return cv2.flip(img, 1)
-
-
-@preserve_shape
-def random_flip(img, code):
-    return cv2.flip(img, code)
-
-
-def transpose(img):
-    return img.transpose(1, 0, 2) if len(img.shape) > 2 else img.transpose(1, 0)
-
-
-def rot90(img, factor):
-    img = np.rot90(img, factor)
-    return np.ascontiguousarray(img)
 
 
 def normalize_cv2(img, mean, denominator):
@@ -686,58 +640,6 @@ def clahe(img, clip_limit=2.0, tile_grid_size=(8, 8)):
     return img
 
 
-@preserve_channel_dim
-def pad(img, min_height, min_width, border_mode=cv2.BORDER_REFLECT_101, value=None):
-    height, width = img.shape[:2]
-
-    if height < min_height:
-        h_pad_top = int((min_height - height) / 2.0)
-        h_pad_bottom = min_height - height - h_pad_top
-    else:
-        h_pad_top = 0
-        h_pad_bottom = 0
-
-    if width < min_width:
-        w_pad_left = int((min_width - width) / 2.0)
-        w_pad_right = min_width - width - w_pad_left
-    else:
-        w_pad_left = 0
-        w_pad_right = 0
-
-    img = pad_with_params(img, h_pad_top, h_pad_bottom, w_pad_left, w_pad_right, border_mode, value)
-
-    if img.shape[:2] != (max(min_height, height), max(min_width, width)):
-        raise RuntimeError(
-            "Invalid result shape. Got: {}. Expected: {}".format(
-                img.shape[:2], (max(min_height, height), max(min_width, width))
-            )
-        )
-
-    return img
-
-
-@preserve_channel_dim
-def pad_with_params(
-    img: np.ndarray,
-    h_pad_top: int,
-    h_pad_bottom: int,
-    w_pad_left: int,
-    w_pad_right: int,
-    border_mode: int = cv2.BORDER_REFLECT_101,
-    value: Optional[int] = None,
-) -> np.ndarray:
-    pad_fn = _maybe_process_in_chunks(
-        cv2.copyMakeBorder,
-        top=h_pad_top,
-        bottom=h_pad_bottom,
-        left=w_pad_left,
-        right=w_pad_right,
-        borderType=border_mode,
-        value=value,
-    )
-    return pad_fn(img)
-
-
 @preserve_shape
 def blur(img, ksize):
     blur_fn = _maybe_process_in_chunks(cv2.blur, ksize=(ksize, ksize))
@@ -1063,185 +965,6 @@ def add_shadow(img, vertices_list):
     return image_rgb
 
 
-@preserve_shape
-def optical_distortion(
-    img,
-    k=0,
-    dx=0,
-    dy=0,
-    interpolation=cv2.INTER_LINEAR,
-    border_mode=cv2.BORDER_REFLECT_101,
-    value=None,
-):
-    """Barrel / pincushion distortion. Unconventional augment.
-
-    Reference:
-        |  https://stackoverflow.com/questions/6199636/formulas-for-barrel-pincushion-distortion
-        |  https://stackoverflow.com/questions/10364201/image-transformation-in-opencv
-        |  https://stackoverflow.com/questions/2477774/correcting-fisheye-distortion-programmatically
-        |  http://www.coldvision.io/2017/03/02/advanced-lane-finding-using-opencv/
-    """
-    height, width = img.shape[:2]
-
-    fx = width
-    fy = height
-
-    cx = width * 0.5 + dx
-    cy = height * 0.5 + dy
-
-    camera_matrix = np.array([[fx, 0, cx], [0, fy, cy], [0, 0, 1]], dtype=np.float32)
-
-    distortion = np.array([k, k, 0, 0, 0], dtype=np.float32)
-    map1, map2 = cv2.initUndistortRectifyMap(camera_matrix, distortion, None, None, (width, height), cv2.CV_32FC1)
-    img = cv2.remap(
-        img,
-        map1,
-        map2,
-        interpolation=interpolation,
-        borderMode=border_mode,
-        borderValue=value,
-    )
-    return img
-
-
-@preserve_shape
-def grid_distortion(
-    img,
-    num_steps=10,
-    xsteps=(),
-    ysteps=(),
-    interpolation=cv2.INTER_LINEAR,
-    border_mode=cv2.BORDER_REFLECT_101,
-    value=None,
-):
-    """Perform a grid distortion of an input image.
-
-    Reference:
-        http://pythology.blogspot.sg/2014/03/interpolation-on-regular-distorted-grid.html
-    """
-    height, width = img.shape[:2]
-
-    x_step = width // num_steps
-    xx = np.zeros(width, np.float32)
-    prev = 0
-    for idx in range(num_steps + 1):
-        x = idx * x_step
-        start = int(x)
-        end = int(x) + x_step
-        if end > width:
-            end = width
-            cur = width
-        else:
-            cur = prev + x_step * xsteps[idx]
-
-        xx[start:end] = np.linspace(prev, cur, end - start)
-        prev = cur
-
-    y_step = height // num_steps
-    yy = np.zeros(height, np.float32)
-    prev = 0
-    for idx in range(num_steps + 1):
-        y = idx * y_step
-        start = int(y)
-        end = int(y) + y_step
-        if end > height:
-            end = height
-            cur = height
-        else:
-            cur = prev + y_step * ysteps[idx]
-
-        yy[start:end] = np.linspace(prev, cur, end - start)
-        prev = cur
-
-    map_x, map_y = np.meshgrid(xx, yy)
-    map_x = map_x.astype(np.float32)
-    map_y = map_y.astype(np.float32)
-
-    remap_fn = _maybe_process_in_chunks(
-        cv2.remap,
-        map1=map_x,
-        map2=map_y,
-        interpolation=interpolation,
-        borderMode=border_mode,
-        borderValue=value,
-    )
-    return remap_fn(img)
-
-
-@preserve_shape
-def elastic_transform_approx(
-    img,
-    alpha,
-    sigma,
-    alpha_affine,
-    interpolation=cv2.INTER_LINEAR,
-    border_mode=cv2.BORDER_REFLECT_101,
-    value=None,
-    random_state=None,
-):
-    """Elastic deformation of images as described in [Simard2003]_ (with modifications for speed).
-    Based on https://gist.github.com/ernestum/601cdf56d2b424757de5
-
-    .. [Simard2003] Simard, Steinkraus and Platt, "Best Practices for
-         Convolutional Neural Networks applied to Visual Document Analysis", in
-         Proc. of the International Conference on Document Analysis and
-         Recognition, 2003.
-    """
-    height, width = img.shape[:2]
-
-    # Random affine
-    center_square = np.float32((height, width)) // 2
-    square_size = min((height, width)) // 3
-    alpha = float(alpha)
-    sigma = float(sigma)
-    alpha_affine = float(alpha_affine)
-
-    pts1 = np.float32(
-        [
-            center_square + square_size,
-            [center_square[0] + square_size, center_square[1] - square_size],
-            center_square - square_size,
-        ]
-    )
-    pts2 = pts1 + random_utils.uniform(-alpha_affine, alpha_affine, size=pts1.shape, random_state=random_state).astype(
-        np.float32
-    )
-    matrix = cv2.getAffineTransform(pts1, pts2)
-
-    warp_fn = _maybe_process_in_chunks(
-        cv2.warpAffine,
-        M=matrix,
-        dsize=(width, height),
-        flags=interpolation,
-        borderMode=border_mode,
-        borderValue=value,
-    )
-    img = warp_fn(img)
-
-    dx = random_utils.rand(height, width, random_state=random_state).astype(np.float32) * 2 - 1
-    cv2.GaussianBlur(dx, (17, 17), sigma, dst=dx)
-    dx *= alpha
-
-    dy = random_utils.rand(height, width, random_state=random_state).astype(np.float32) * 2 - 1
-    cv2.GaussianBlur(dy, (17, 17), sigma, dst=dy)
-    dy *= alpha
-
-    x, y = np.meshgrid(np.arange(width), np.arange(height))
-
-    map_x = np.float32(x + dx)
-    map_y = np.float32(y + dy)
-
-    remap_fn = _maybe_process_in_chunks(
-        cv2.remap,
-        map1=map_x,
-        map2=map_y,
-        interpolation=interpolation,
-        borderMode=border_mode,
-        borderValue=value,
-    )
-    return remap_fn(img)
-
-
 def invert(img):
     return 255 - img
 
@@ -1397,159 +1120,6 @@ def from_float(img, dtype, max_value=None):
     return (img * max_value).astype(dtype)
 
 
-def bbox_vflip(bbox, rows, cols):  # skipcq: PYL-W0613
-    """Flip a bounding box vertically around the x-axis.
-
-    Args:
-        bbox (tuple): A bounding box `(x_min, y_min, x_max, y_max)`.
-        rows (int): Image rows.
-        cols (int): Image cols.
-
-    Returns:
-        tuple: A bounding box `(x_min, y_min, x_max, y_max)`.
-
-    """
-    x_min, y_min, x_max, y_max = bbox[:4]
-    return x_min, 1 - y_max, x_max, 1 - y_min
-
-
-def bbox_hflip(bbox, rows, cols):  # skipcq: PYL-W0613
-    """Flip a bounding box horizontally around the y-axis.
-
-    Args:
-        bbox (tuple): A bounding box `(x_min, y_min, x_max, y_max)`.
-        rows (int): Image rows.
-        cols (int): Image cols.
-
-    Returns:
-        tuple: A bounding box `(x_min, y_min, x_max, y_max)`.
-
-    """
-    x_min, y_min, x_max, y_max = bbox[:4]
-    return 1 - x_max, y_min, 1 - x_min, y_max
-
-
-def bbox_flip(bbox, d, rows, cols):
-    """Flip a bounding box either vertically, horizontally or both depending on the value of `d`.
-
-    Args:
-        bbox (tuple): A bounding box `(x_min, y_min, x_max, y_max)`.
-        d (int):
-        rows (int): Image rows.
-        cols (int): Image cols.
-
-    Returns:
-        tuple: A bounding box `(x_min, y_min, x_max, y_max)`.
-
-    Raises:
-        ValueError: if value of `d` is not -1, 0 or 1.
-
-    """
-    if d == 0:
-        bbox = bbox_vflip(bbox, rows, cols)
-    elif d == 1:
-        bbox = bbox_hflip(bbox, rows, cols)
-    elif d == -1:
-        bbox = bbox_hflip(bbox, rows, cols)
-        bbox = bbox_vflip(bbox, rows, cols)
-    else:
-        raise ValueError("Invalid d value {}. Valid values are -1, 0 and 1".format(d))
-    return bbox
-
-
-def bbox_transpose(bbox, axis, rows, cols):  # skipcq: PYL-W0613
-    """Transposes a bounding box along given axis.
-
-    Args:
-        bbox (tuple): A bounding box `(x_min, y_min, x_max, y_max)`.
-        axis (int): 0 - main axis, 1 - secondary axis.
-        rows (int): Image rows.
-        cols (int): Image cols.
-
-    Returns:
-        tuple: A bounding box tuple `(x_min, y_min, x_max, y_max)`.
-
-    Raises:
-        ValueError: If axis not equal to 0 or 1.
-
-    """
-    x_min, y_min, x_max, y_max = bbox[:4]
-    if axis not in {0, 1}:
-        raise ValueError("Axis must be either 0 or 1.")
-    if axis == 0:
-        bbox = (y_min, x_min, y_max, x_max)
-    if axis == 1:
-        bbox = (1 - y_max, 1 - x_max, 1 - y_min, 1 - x_min)
-    return bbox
-
-
-@angle_2pi_range
-def keypoint_vflip(keypoint, rows, cols):
-    """Flip a keypoint vertically around the x-axis.
-
-    Args:
-        keypoint (tuple): A keypoint `(x, y, angle, scale)`.
-        rows (int): Image height.
-        cols( int): Image width.
-
-    Returns:
-        tuple: A keypoint `(x, y, angle, scale)`.
-
-    """
-    x, y, angle, scale = keypoint
-    angle = -angle
-    return x, (rows - 1) - y, angle, scale
-
-
-@angle_2pi_range
-def keypoint_hflip(keypoint, rows, cols):
-    """Flip a keypoint horizontally around the y-axis.
-
-    Args:
-        keypoint (tuple): A keypoint `(x, y, angle, scale)`.
-        rows (int): Image height.
-        cols (int): Image width.
-
-    Returns:
-        tuple: A keypoint `(x, y, angle, scale)`.
-
-    """
-    x, y, angle, scale = keypoint
-    angle = math.pi - angle
-    return (cols - 1) - x, y, angle, scale
-
-
-def keypoint_flip(keypoint, d, rows, cols):
-    """Flip a keypoint either vertically, horizontally or both depending on the value of `d`.
-
-    Args:
-        keypoint (tuple): A keypoint `(x, y, angle, scale)`.
-        d (int): Number of flip. Must be -1, 0 or 1:
-            * 0 - vertical flip,
-            * 1 - horizontal flip,
-            * -1 - vertical and horizontal flip.
-        rows (int): Image height.
-        cols (int): Image width.
-
-    Returns:
-        tuple: A keypoint `(x, y, angle, scale)`.
-
-    Raises:
-        ValueError: if value of `d` is not -1, 0 or 1.
-
-    """
-    if d == 0:
-        keypoint = keypoint_vflip(keypoint, rows, cols)
-    elif d == 1:
-        keypoint = keypoint_hflip(keypoint, rows, cols)
-    elif d == -1:
-        keypoint = keypoint_hflip(keypoint, rows, cols)
-        keypoint = keypoint_vflip(keypoint, rows, cols)
-    else:
-        raise ValueError("Invalid d value {}. Valid values are -1, 0 and 1".format(d))
-    return keypoint
-
-
 def noop(input_obj, **params):  # skipcq: PYL-W0613
     return input_obj
 
@@ -1577,26 +1147,6 @@ def swap_tiles_on_image(image, tiles):
         ]
 
     return new_image
-
-
-def keypoint_transpose(keypoint):
-    """Rotate a keypoint by angle.
-
-    Args:
-        keypoint (tuple): A keypoint `(x, y, angle, scale)`.
-
-    Returns:
-        tuple: A keypoint `(x, y, angle, scale)`.
-
-    """
-    x, y, angle, scale = keypoint[:4]
-
-    if angle <= np.pi:
-        angle = np.pi - angle
-    else:
-        angle = 3 * np.pi - angle
-
-    return y, x, angle, scale
 
 
 @clipped
