+++ albumentations/augmentations/geometric/functional.py
@@ -19,6 +19,11 @@ from ..functional import (
 )
 
 __all__ = [
+    "optical_distortion",
+    "elastic_transform_approx",
+    "grid_distortion",
+    "pad",
+    "pad_with_params",
     "bbox_rot90",
     "keypoint_rot90",
     "rotate",
@@ -51,17 +56,28 @@ __all__ = [
     "from_distance_maps",
     "keypoint_piecewise_affine",
     "bbox_piecewise_affine",
+    "bbox_flip",
+    "bbox_hflip",
+    "bbox_transpose",
+    "bbox_vflip",
+    "hflip",
+    "hflip_cv2",
+    "transpose",
+    "keypoint_flip",
+    "keypoint_hflip",
+    "keypoint_transpose",
+    "keypoint_vflip",
 ]
 
 
-def bbox_rot90(bbox, factor, rows, cols):  # skipcq: PYL-W0613
+def bbox_rot90(bbox: BoxType, factor: int, rows: int, cols: int) -> BoxType:  # skipcq: PYL-W0613
     """Rotates a bounding box by 90 degrees CCW (see np.rot90)
 
     Args:
-        bbox (tuple): A bounding box tuple (x_min, y_min, x_max, y_max).
-        factor (int): Number of CCW rotations. Must be in set {0, 1, 2, 3} See np.rot90.
-        rows (int): Image rows.
-        cols (int): Image cols.
+        bbox: A bounding box tuple (x_min, y_min, x_max, y_max).
+        factor: Number of CCW rotations. Must be in set {0, 1, 2, 3} See np.rot90.
+        rows: Image rows.
+        cols: Image cols.
 
     Returns:
         tuple: A bounding box tuple (x_min, y_min, x_max, y_max).
@@ -80,14 +96,14 @@ def bbox_rot90(bbox, factor, rows, cols):  # skipcq: PYL-W0613
 
 
 @angle_2pi_range
-def keypoint_rot90(keypoint, factor, rows, cols, **params):
+def keypoint_rot90(keypoint: KeypointType, factor: int, rows: int, cols: int, **params) -> KeypointType:
     """Rotates a keypoint by 90 degrees CCW (see np.rot90)
 
     Args:
-        keypoint (tuple): A keypoint `(x, y, angle, scale)`.
-        factor (int): Number of CCW rotations. Must be in range [0;3] See np.rot90.
-        rows (int): Image height.
-        cols (int): Image width.
+        keypoint: A keypoint `(x, y, angle, scale)`.
+        factor: Number of CCW rotations. Must be in range [0;3] See np.rot90.
+        rows: Image height.
+        cols: Image width.
 
     Returns:
         tuple: A keypoint `(x, y, angle, scale)`.
@@ -112,7 +128,13 @@ def keypoint_rot90(keypoint, factor, rows, cols, **params):
 
 
 @preserve_channel_dim
-def rotate(img, angle, interpolation=cv2.INTER_LINEAR, border_mode=cv2.BORDER_REFLECT_101, value=None):
+def rotate(
+    img: np.ndarray,
+    angle: int,
+    interpolation: int = cv2.INTER_LINEAR,
+    border_mode: int = cv2.BORDER_REFLECT_101,
+    value: Optional[Tuple[int, ...]] = None,
+):
     height, width = img.shape[:2]
     # for images we use additional shifts of (0.5, 0.5) as otherwise
     # we get an ugly black border for 90deg rotations
@@ -124,15 +146,15 @@ def rotate(img, angle, interpolation=cv2.INTER_LINEAR, border_mode=cv2.BORDER_RE
     return warp_fn(img)
 
 
-def bbox_rotate(bbox, angle, method, rows, cols):
+def bbox_rotate(bbox: BoxType, angle: int, method: str, rows: int, cols: int) -> BoxType:
     """Rotates a bounding box by angle degrees.
 
     Args:
-        bbox (tuple): A bounding box `(x_min, y_min, x_max, y_max)`.
-        angle (int): Angle of rotation in degrees.
-        method(str): Rotation method used. Should be one of: "largest_box", "ellipse". Default: "largest_box".
-        rows (int): Image rows.
-        cols (int): Image cols.
+        bbox: A bounding box `(x_min, y_min, x_max, y_max)`.
+        angle: Angle of rotation in degrees.
+        method: Rotation method used. Should be one of: "largest_box", "ellipse". Default: "largest_box".
+        rows: Image rows.
+        cols: Image cols.
 
     Returns:
         A bounding box `(x_min, y_min, x_max, y_max)`.
@@ -365,17 +387,17 @@ def resize(img, height, width, interpolation=cv2.INTER_LINEAR):
 
 
 @preserve_channel_dim
-def scale(img, scale, interpolation=cv2.INTER_LINEAR):
+def scale(img: np.ndarray, scale: float, interpolation: int = cv2.INTER_LINEAR) -> np.ndarray:
     height, width = img.shape[:2]
     new_height, new_width = int(height * scale), int(width * scale)
     return resize(img, new_height, new_width, interpolation)
 
 
-def keypoint_scale(keypoint: Sequence[float], scale_x: float, scale_y: float):
+def keypoint_scale(keypoint: KeypointType, scale_x: float, scale_y: float) -> KeypointType:
     """Scales a keypoint by scale_x and scale_y.
 
     Args:
-        keypoint (tuple): A keypoint `(x, y, angle, scale)`.
+        keypoint: A keypoint `(x, y, angle, scale)`.
         scale_x: Scale coefficient x-axis.
         scale_y: Scale coefficient y-axis.
 
@@ -407,12 +429,12 @@ def _func_max_size(img, max_size, interpolation, func):
 
 
 @preserve_channel_dim
-def longest_max_size(img, max_size, interpolation):
+def longest_max_size(img: np.ndarray, max_size: int, interpolation: int) -> np.ndarray:
     return _func_max_size(img, max_size, interpolation, max)
 
 
 @preserve_channel_dim
-def smallest_max_size(img, max_size, interpolation):
+def smallest_max_size(img: np.ndarray, max_size: int, interpolation: int) -> np.ndarray:
     return _func_max_size(img, max_size, interpolation, min)
 
 
@@ -445,15 +467,15 @@ def perspective(
 
 
 def perspective_bbox(
-    bbox: Sequence[float],
+    bbox: BoxType,
     height: int,
     width: int,
     matrix: np.ndarray,
     max_width: int,
     max_height: int,
     keep_size: bool,
-):
-    x1, y1, x2, y2 = denormalize_bbox(bbox, height, width)
+) -> BoxType:
+    x1, y1, x2, y2 = denormalize_bbox(bbox, height, width)[:4]
 
     points = np.array([[x1, y1], [x2, y1], [x2, y2], [x1, y2]], dtype=np.float32)
 
@@ -559,7 +581,7 @@ def bbox_affine(
     if _is_identity_matrix(matrix):
         return bbox
 
-    x_min, y_min, x_max, y_max = denormalize_bbox(bbox, rows, cols)
+    x_min, y_min, x_max, y_max = denormalize_bbox(bbox, rows, cols)[:4]
     points = np.array(
         [
             [x_min, y_min],
@@ -598,7 +620,7 @@ def safe_rotate(
 
 
 def bbox_safe_rotate(bbox: BoxType, matrix: np.ndarray, cols: int, rows: int) -> BoxType:
-    x1, y1, x2, y2 = denormalize_bbox(bbox, rows, cols)
+    x1, y1, x2, y2 = denormalize_bbox(bbox, rows, cols)[:4]
     points = np.array(
         [
             [x1, y1, 1],
@@ -636,7 +658,7 @@ def keypoint_safe_rotate(
     cols: int,
     rows: int,
 ) -> KeypointType:
-    x, y, a, s = keypoint
+    x, y, a, s = keypoint[:4]
     point = np.array([[x, y, 1]])
     x, y = (point @ matrix.T)[0]
 
@@ -663,7 +685,7 @@ def piecewise_affine(
 
 
 def to_distance_maps(
-    keypoints: Sequence[Sequence[float]], height: int, width: int, inverted: bool = False
+    keypoints: Sequence[Tuple[float, float]], height: int, width: int, inverted: bool = False
 ) -> np.ndarray:
     """Generate a ``(H,W,N)`` array of distance maps for ``N`` keypoints.
 
@@ -674,16 +696,16 @@ def to_distance_maps(
     method that only supports the augmentation of images.
 
     Args:
-        keypoint (sequence of float): keypoint coordinates
-        height (int): image height
-        width (int): image width
+        keypoint: keypoint coordinates
+        height: image height
+        width: image width
         inverted (bool): If ``True``, inverted distance maps are returned where each
             distance value d is replaced by ``d/(d+1)``, i.e. the distance
             maps have values in the range ``(0.0, 1.0]`` with ``1.0`` denoting
             exactly the position of the respective keypoint.
 
     Returns:
-        (H,W,N) ndarray
+        (H, W, N) ndarray
             A ``float32`` array containing ``N`` distance maps for ``N``
             keypoints. Each location ``(y, x, n)`` in the array denotes the
             euclidean distance at ``(y, x)`` to the ``n``-th keypoint.
@@ -784,13 +806,13 @@ def from_distance_maps(
 
 
 def keypoint_piecewise_affine(
-    keypoint: Sequence[float],
+    keypoint: KeypointType,
     matrix: skimage.transform.PiecewiseAffineTransform,
     h: int,
     w: int,
     keypoints_threshold: float,
-) -> Tuple[float, float, float, float]:
-    x, y, a, s = keypoint
+) -> KeypointType:
+    x, y, a, s = keypoint[:4]
     dist_maps = to_distance_maps([(x, y)], h, w, True)
     dist_maps = piecewise_affine(dist_maps, matrix, 0, "constant", 0)
     x, y = from_distance_maps(dist_maps, True, {"x": -1, "y": -1}, keypoints_threshold)[0]
@@ -798,13 +820,13 @@ def keypoint_piecewise_affine(
 
 
 def bbox_piecewise_affine(
-    bbox: Sequence[float],
+    bbox: BoxType,
     matrix: skimage.transform.PiecewiseAffineTransform,
     h: int,
     w: int,
     keypoints_threshold: float,
-) -> Tuple[float, float, float, float]:
-    x1, y1, x2, y2 = denormalize_bbox(tuple(bbox), h, w)
+) -> BoxType:
+    x1, y1, x2, y2 = denormalize_bbox(bbox, h, w)[:4]
     keypoints = [
         (x1, y1),
         (x2, y1),
@@ -821,3 +843,439 @@ def bbox_piecewise_affine(
     x2 = keypoints_arr[:, 0].max()
     y2 = keypoints_arr[:, 1].max()
     return typing.cast(BoxType, normalize_bbox((x1, y1, x2, y2), h, w))
+
+
+def vflip(img: np.ndarray) -> np.ndarray:
+    return np.ascontiguousarray(img[::-1, ...])
+
+
+def hflip(img: np.ndarray) -> np.ndarray:
+    return np.ascontiguousarray(img[:, ::-1, ...])
+
+
+def hflip_cv2(img: np.ndarray) -> np.ndarray:
+    return cv2.flip(img, 1)
+
+
+@preserve_shape
+def random_flip(img, code):
+    return cv2.flip(img, code)
+
+
+def transpose(img: np.ndarray) -> np.ndarray:
+    return img.transpose(1, 0, 2) if len(img.shape) > 2 else img.transpose(1, 0)
+
+
+def rot90(img: np.ndarray, factor: int) -> np.ndarray:
+    img = np.rot90(img, factor)
+    return np.ascontiguousarray(img)
+
+
+def bbox_vflip(bbox: BoxType, rows: int, cols: int) -> BoxType:  # skipcq: PYL-W0613
+    """Flip a bounding box vertically around the x-axis.
+
+    Args:
+        bbox: A bounding box `(x_min, y_min, x_max, y_max)`.
+        rows: Image rows.
+        cols: Image cols.
+
+    Returns:
+        tuple: A bounding box `(x_min, y_min, x_max, y_max)`.
+
+    """
+    x_min, y_min, x_max, y_max = bbox[:4]
+    return x_min, 1 - y_max, x_max, 1 - y_min
+
+
+def bbox_hflip(bbox: BoxType, rows: int, cols: int) -> BoxType:  # skipcq: PYL-W0613
+    """Flip a bounding box horizontally around the y-axis.
+
+    Args:
+        bbox: A bounding box `(x_min, y_min, x_max, y_max)`.
+        rows: Image rows.
+        cols: Image cols.
+
+    Returns:
+        A bounding box `(x_min, y_min, x_max, y_max)`.
+
+    """
+    x_min, y_min, x_max, y_max = bbox[:4]
+    return 1 - x_max, y_min, 1 - x_min, y_max
+
+
+def bbox_flip(bbox: BoxType, d: int, rows: int, cols: int) -> BoxType:
+    """Flip a bounding box either vertically, horizontally or both depending on the value of `d`.
+
+    Args:
+        bbox: A bounding box `(x_min, y_min, x_max, y_max)`.
+        d: dimension. 0 for vertical flip, 1 for horizontal, -1 for transpose
+        rows: Image rows.
+        cols: Image cols.
+
+    Returns:
+        A bounding box `(x_min, y_min, x_max, y_max)`.
+
+    Raises:
+        ValueError: if value of `d` is not -1, 0 or 1.
+
+    """
+    if d == 0:
+        bbox = bbox_vflip(bbox, rows, cols)
+    elif d == 1:
+        bbox = bbox_hflip(bbox, rows, cols)
+    elif d == -1:
+        bbox = bbox_hflip(bbox, rows, cols)
+        bbox = bbox_vflip(bbox, rows, cols)
+    else:
+        raise ValueError("Invalid d value {}. Valid values are -1, 0 and 1".format(d))
+    return bbox
+
+
+def bbox_transpose(bbox: KeypointType, axis: int, rows: int, cols: int) -> KeypointType:  # skipcq: PYL-W0613
+    """Transposes a bounding box along given axis.
+
+    Args:
+        bbox: A bounding box `(x_min, y_min, x_max, y_max)`.
+        axis: 0 - main axis, 1 - secondary axis.
+        rows: Image rows.
+        cols: Image cols.
+
+    Returns:
+        A bounding box tuple `(x_min, y_min, x_max, y_max)`.
+
+    Raises:
+        ValueError: If axis not equal to 0 or 1.
+
+    """
+    x_min, y_min, x_max, y_max = bbox[:4]
+    if axis not in {0, 1}:
+        raise ValueError("Axis must be either 0 or 1.")
+    if axis == 0:
+        bbox = (y_min, x_min, y_max, x_max)
+    if axis == 1:
+        bbox = (1 - y_max, 1 - x_max, 1 - y_min, 1 - x_min)
+    return bbox
+
+
+@angle_2pi_range
+def keypoint_vflip(keypoint: KeypointType, rows: int, cols: int) -> KeypointType:
+    """Flip a keypoint vertically around the x-axis.
+
+    Args:
+        keypoint: A keypoint `(x, y, angle, scale)`.
+        rows: Image height.
+        cols: Image width.
+
+    Returns:
+        tuple: A keypoint `(x, y, angle, scale)`.
+
+    """
+    x, y, angle, scale = keypoint[:4]
+    angle = -angle
+    return x, (rows - 1) - y, angle, scale
+
+
+@angle_2pi_range
+def keypoint_hflip(keypoint: KeypointType, rows: int, cols: int) -> KeypointType:
+    """Flip a keypoint horizontally around the y-axis.
+
+    Args:
+        keypoint: A keypoint `(x, y, angle, scale)`.
+        rows: Image height.
+        cols: Image width.
+
+    Returns:
+        A keypoint `(x, y, angle, scale)`.
+
+    """
+    x, y, angle, scale = keypoint[:4]
+    angle = math.pi - angle
+    return (cols - 1) - x, y, angle, scale
+
+
+def keypoint_flip(keypoint: KeypointType, d: int, rows: int, cols: int) -> KeypointType:
+    """Flip a keypoint either vertically, horizontally or both depending on the value of `d`.
+
+    Args:
+        keypoint: A keypoint `(x, y, angle, scale)`.
+        d: Number of flip. Must be -1, 0 or 1:
+            * 0 - vertical flip,
+            * 1 - horizontal flip,
+            * -1 - vertical and horizontal flip.
+        rows: Image height.
+        cols: Image width.
+
+    Returns:
+        A keypoint `(x, y, angle, scale)`.
+
+    Raises:
+        ValueError: if value of `d` is not -1, 0 or 1.
+
+    """
+    if d == 0:
+        keypoint = keypoint_vflip(keypoint, rows, cols)
+    elif d == 1:
+        keypoint = keypoint_hflip(keypoint, rows, cols)
+    elif d == -1:
+        keypoint = keypoint_hflip(keypoint, rows, cols)
+        keypoint = keypoint_vflip(keypoint, rows, cols)
+    else:
+        raise ValueError(f"Invalid d value {d}. Valid values are -1, 0 and 1")
+    return keypoint
+
+
+def keypoint_transpose(keypoint: KeypointType) -> KeypointType:
+    """Rotate a keypoint by angle.
+
+    Args:
+        keypoint: A keypoint `(x, y, angle, scale)`.
+
+    Returns:
+        A keypoint `(x, y, angle, scale)`.
+
+    """
+    x, y, angle, scale = keypoint[:4]
+
+    if angle <= np.pi:
+        angle = np.pi - angle
+    else:
+        angle = 3 * np.pi - angle
+
+    return y, x, angle, scale
+
+
+@preserve_channel_dim
+def pad(
+    img: np.ndarray,
+    min_height: int,
+    min_width: int,
+    border_mode: int = cv2.BORDER_REFLECT_101,
+    value: Optional[Tuple[int, ...]] = None,
+) -> np.ndarray:
+    height, width = img.shape[:2]
+
+    if height < min_height:
+        h_pad_top = int((min_height - height) / 2.0)
+        h_pad_bottom = min_height - height - h_pad_top
+    else:
+        h_pad_top = 0
+        h_pad_bottom = 0
+
+    if width < min_width:
+        w_pad_left = int((min_width - width) / 2.0)
+        w_pad_right = min_width - width - w_pad_left
+    else:
+        w_pad_left = 0
+        w_pad_right = 0
+
+    img = pad_with_params(img, h_pad_top, h_pad_bottom, w_pad_left, w_pad_right, border_mode, value)
+
+    if img.shape[:2] != (max(min_height, height), max(min_width, width)):
+        raise RuntimeError(
+            "Invalid result shape. Got: {}. Expected: {}".format(
+                img.shape[:2], (max(min_height, height), max(min_width, width))
+            )
+        )
+
+    return img
+
+
+@preserve_channel_dim
+def pad_with_params(
+    img: np.ndarray,
+    h_pad_top: int,
+    h_pad_bottom: int,
+    w_pad_left: int,
+    w_pad_right: int,
+    border_mode: int = cv2.BORDER_REFLECT_101,
+    value: Optional[int] = None,
+) -> np.ndarray:
+    pad_fn = _maybe_process_in_chunks(
+        cv2.copyMakeBorder,
+        top=h_pad_top,
+        bottom=h_pad_bottom,
+        left=w_pad_left,
+        right=w_pad_right,
+        borderType=border_mode,
+        value=value,
+    )
+    return pad_fn(img)
+
+
+@preserve_shape
+def optical_distortion(
+    img,
+    k=0,
+    dx=0,
+    dy=0,
+    interpolation=cv2.INTER_LINEAR,
+    border_mode=cv2.BORDER_REFLECT_101,
+    value=None,
+):
+    """Barrel / pincushion distortion. Unconventional augment.
+
+    Reference:
+        |  https://stackoverflow.com/questions/6199636/formulas-for-barrel-pincushion-distortion
+        |  https://stackoverflow.com/questions/10364201/image-transformation-in-opencv
+        |  https://stackoverflow.com/questions/2477774/correcting-fisheye-distortion-programmatically
+        |  http://www.coldvision.io/2017/03/02/advanced-lane-finding-using-opencv/
+    """
+    height, width = img.shape[:2]
+
+    fx = width
+    fy = height
+
+    cx = width * 0.5 + dx
+    cy = height * 0.5 + dy
+
+    camera_matrix = np.array([[fx, 0, cx], [0, fy, cy], [0, 0, 1]], dtype=np.float32)
+
+    distortion = np.array([k, k, 0, 0, 0], dtype=np.float32)
+    map1, map2 = cv2.initUndistortRectifyMap(camera_matrix, distortion, None, None, (width, height), cv2.CV_32FC1)
+    img = cv2.remap(
+        img,
+        map1,
+        map2,
+        interpolation=interpolation,
+        borderMode=border_mode,
+        borderValue=value,
+    )
+    return img
+
+
+@preserve_shape
+def grid_distortion(
+    img,
+    num_steps=10,
+    xsteps=(),
+    ysteps=(),
+    interpolation=cv2.INTER_LINEAR,
+    border_mode=cv2.BORDER_REFLECT_101,
+    value=None,
+):
+    """Perform a grid distortion of an input image.
+
+    Reference:
+        http://pythology.blogspot.sg/2014/03/interpolation-on-regular-distorted-grid.html
+    """
+    height, width = img.shape[:2]
+
+    x_step = width // num_steps
+    xx = np.zeros(width, np.float32)
+    prev = 0
+    for idx in range(num_steps + 1):
+        x = idx * x_step
+        start = int(x)
+        end = int(x) + x_step
+        if end > width:
+            end = width
+            cur = width
+        else:
+            cur = prev + x_step * xsteps[idx]
+
+        xx[start:end] = np.linspace(prev, cur, end - start)
+        prev = cur
+
+    y_step = height // num_steps
+    yy = np.zeros(height, np.float32)
+    prev = 0
+    for idx in range(num_steps + 1):
+        y = idx * y_step
+        start = int(y)
+        end = int(y) + y_step
+        if end > height:
+            end = height
+            cur = height
+        else:
+            cur = prev + y_step * ysteps[idx]
+
+        yy[start:end] = np.linspace(prev, cur, end - start)
+        prev = cur
+
+    map_x, map_y = np.meshgrid(xx, yy)
+    map_x = map_x.astype(np.float32)
+    map_y = map_y.astype(np.float32)
+
+    remap_fn = _maybe_process_in_chunks(
+        cv2.remap,
+        map1=map_x,
+        map2=map_y,
+        interpolation=interpolation,
+        borderMode=border_mode,
+        borderValue=value,
+    )
+    return remap_fn(img)
+
+
+@preserve_shape
+def elastic_transform_approx(
+    img,
+    alpha,
+    sigma,
+    alpha_affine,
+    interpolation=cv2.INTER_LINEAR,
+    border_mode=cv2.BORDER_REFLECT_101,
+    value=None,
+    random_state=None,
+):
+    """Elastic deformation of images as described in [Simard2003]_ (with modifications for speed).
+    Based on https://gist.github.com/ernestum/601cdf56d2b424757de5
+
+    .. [Simard2003] Simard, Steinkraus and Platt, "Best Practices for
+         Convolutional Neural Networks applied to Visual Document Analysis", in
+         Proc. of the International Conference on Document Analysis and
+         Recognition, 2003.
+    """
+    height, width = img.shape[:2]
+
+    # Random affine
+    center_square = np.float32((height, width)) // 2
+    square_size = min((height, width)) // 3
+    alpha = float(alpha)
+    sigma = float(sigma)
+    alpha_affine = float(alpha_affine)
+
+    pts1 = np.float32(
+        [
+            center_square + square_size,
+            [center_square[0] + square_size, center_square[1] - square_size],
+            center_square - square_size,
+        ]
+    )
+    pts2 = pts1 + random_utils.uniform(-alpha_affine, alpha_affine, size=pts1.shape, random_state=random_state).astype(
+        np.float32
+    )
+    matrix = cv2.getAffineTransform(pts1, pts2)
+
+    warp_fn = _maybe_process_in_chunks(
+        cv2.warpAffine,
+        M=matrix,
+        dsize=(width, height),
+        flags=interpolation,
+        borderMode=border_mode,
+        borderValue=value,
+    )
+    img = warp_fn(img)
+
+    dx = random_utils.rand(height, width, random_state=random_state).astype(np.float32) * 2 - 1
+    cv2.GaussianBlur(dx, (17, 17), sigma, dst=dx)
+    dx *= alpha
+
+    dy = random_utils.rand(height, width, random_state=random_state).astype(np.float32) * 2 - 1
+    cv2.GaussianBlur(dy, (17, 17), sigma, dst=dy)
+    dy *= alpha
+
+    x, y = np.meshgrid(np.arange(width), np.arange(height))
+
+    map_x = np.float32(x + dx)
+    map_y = np.float32(y + dy)
+
+    remap_fn = _maybe_process_in_chunks(
+        cv2.remap,
+        map1=map_x,
+        map2=map_y,
+        interpolation=interpolation,
+        borderMode=border_mode,
+        borderValue=value,
+    )
+    return remap_fn(img)
