+++ albumentations/core/transforms_interface.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import
 
 import random
 from copy import deepcopy
-from typing import Any, Callable, Dict, List, Optional, Sequence, Tuple, Union
+from typing import Any, Callable, Dict, List, Sequence, Tuple, Union, cast
 from warnings import warn
 
 import cv2
@@ -11,11 +11,11 @@ import numpy as np
 from .serialization import Serializable, get_shortest_class_fullname
 from .utils import format_args
 
-__all__ = ["to_tuple", "BasicTransform", "DualTransform", "ImageOnlyTransform", "NoOp"]
+__all__ = ["to_tuple", "BasicTransform", "DualTransform", "ImageOnlyTransform", "NoOp", "BoxType", "KeypointType"]
 
 NumType = Union[int, float, np.ndarray]
-BoxType = Tuple[float, float, float, float]
-KeypointType = Tuple[float, float, float, float]
+BoxType = Union[Tuple[float, float, float, float], Tuple[float, float, float, float, Any]]
+KeypointType = Union[Tuple[float, float, float, float], Tuple[float, float, float, float, Any]]
 
 
 def to_tuple(param, low=None, bias=None):
@@ -227,14 +227,21 @@ class DualTransform(BasicTransform):
     def apply_to_bbox(self, bbox: BoxType, **params) -> BoxType:
         raise NotImplementedError("Method apply_to_bbox is not implemented in class " + self.__class__.__name__)
 
-    def apply_to_keypoint(self, keypoint: KeypointType, **params) -> KeypointType:
+    def apply_to_keypoint(self, keypoint: Tuple[float, float, float, float], **params) -> KeypointType:
         raise NotImplementedError("Method apply_to_keypoint is not implemented in class " + self.__class__.__name__)
 
     def apply_to_bboxes(self, bboxes: Sequence[BoxType], **params) -> List[BoxType]:
         return [self.apply_to_bbox(tuple(bbox[:4]), **params) + tuple(bbox[4:]) for bbox in bboxes]  # type: ignore
 
     def apply_to_keypoints(self, keypoints: Sequence[KeypointType], **params) -> List[KeypointType]:
-        return [self.apply_to_keypoint(tuple(keypoint[:4]), **params) + tuple(keypoint[4:]) for keypoint in keypoints]  # type: ignore # noqa
+        return [
+            cast(
+                KeypointType,
+                self.apply_to_keypoint(cast(Tuple[float, float, float, float], tuple(keypoint[:4])), **params)
+                + tuple(keypoint[4:]),
+            )
+            for keypoint in keypoints
+        ]  # type: ignore # noqa
 
     def apply_to_mask(self, img: np.ndarray, **params) -> np.ndarray:
         return self.apply(img, **{k: cv2.INTER_NEAREST if k == "interpolation" else v for k, v in params.items()})
